Improvements and Optimizations
Better to have fitness be part of the SudokueGrid so we don't have to recalculate everytime we want to get the fitness.

Grids are initialized smarter, making sure there are no duplicates in rows at the least.

Added re-seeding ability to re-seed population if it grows stale. 
Added Dynamic Mutation but it makes the process much slower and IDK if it is worth it yet.